move state functions to state struct methods
